// Ui/src/app/application/common/analyze-options/enums/analyze-option-enum.ts
export enum AnalyzeOptionEnum {
    AvgPathLength = "AvgPathLength",
    Diameter = "Diameter",
    AvgDegree = "AvgDegree",
    AvgClusteringCoefficient = "AvgClusteringCoefficient",
    Cycles3 = "Cycles3",
    Cycles4 = "Cycles4",
    EigenValues = "EigenValues",
    EigenDistanceDistribution = "EigenDistanceDistribution",
    LaplacianEigenValues = "LaplacianEigenValues",
    DegreeDistribution = "DegreeDistribution",
    ClusteringCoefficientDistribution = "ClusteringCoefficientDistribution",
    ClusteringCoefficientPerVertex = "ClusteringCoefficientPerVertex",
    ConnectedComponentDistribution = "ConnectedComponentDistribution",
    DistanceDistribution = "DistanceDistribution",
    TriangleByVertexDistribution = "TriangleByVertexDistribution",
    DegreeCentrality = "DegreeCentrality",
    BetweennessCentrality = "BetweennessCentrality",
    ClosenessCentrality = "ClosenessCentrality",
    Cycles3Evolution = "Cycles3Evolution",
    "Algorithm_1_By_All_Nodes" = "Algorithm_1_By_All_Nodes",
    "Algorithm_2_By_Active_Nodes_List" = "Algorithm_2_By_Active_Nodes_List",
    "Algorithm_Final" = "Algorithm_Final"
}